fixed some warnings on Windows
git-svn-id: http://find-object.googlecode.com/svn/trunk/find_object@425 620bd6b2-0a58-f614-fd9a-1bd335dccda9

+++ src/ObjSignature.h
@@ -93,7 +93,7 @@ public:
 
 		std::vector<unsigned char> bytes;
 		cv::imencode(".png", image_, bytes);
-		streamPtr << QByteArray((char*)bytes.data(), bytes.size());
+		streamPtr << QByteArray((char*)bytes.data(), (int)bytes.size());
 	}
 
 	void load(QDataStream & streamPtr)
